+++ in_app_review/lib/in_app_review.dart
@@ -12,7 +12,7 @@ class InAppReview {
   /// On Android the Google Play Store must be installed and the device must be
   /// running **Android 5 Lollipop(API 21)** or higher.
   ///
-  /// IOS devices must be running **IOS version 10.3** or higher.
+  /// iOS devices must be running **iOS version 10.3** or higher.
   ///
   /// MacOS devices must be running **MacOS version 10.14** or higher
   Future<bool> isAvailable() => InAppReviewPlatform.instance.isAvailable();
@@ -20,8 +20,8 @@ class InAppReview {
   /// Attempts to show the review dialog. It's recommended to first check if
   /// the device supports this feature via [isAvailable].
   ///
-  /// To improve the users experience, IOS and Android enforce limitations
-  /// that might prevent this from working after a few tries. IOS & MacOS users
+  /// To improve the users experience, iOS and Android enforce limitations
+  /// that might prevent this from working after a few tries. iOS & MacOS users
   /// can also disable this feature entirely in the App Store settings.
   ///
   /// More info and guidance:
@@ -31,13 +31,13 @@ class InAppReview {
   Future<void> requestReview() => InAppReviewPlatform.instance.requestReview();
 
   /// Opens the Play Store on Android, the App Store with a review
-  /// screen on IOS & MacOS and the Microsoft Store on Windows.
+  /// screen on iOS & MacOS and the Microsoft Store on Windows.
   ///
-  /// [appStoreId] is required for IOS & MacOS.
+  /// [appStoreId] is required for iOS & MacOS.
   ///
   /// [microsoftStoreId] is required for Windows.
   Future<void> openStoreListing({
-    /// Required for IOS & MacOS.
+    /// Required for iOS & MacOS.
     String? appStoreId,
 
     /// Required for Windows.