+++ in_app_review_platform_interface/lib/in_app_review_platform_interface.dart
@@ -30,7 +30,7 @@ abstract class InAppReviewPlatform extends PlatformInterface {
   /// On Android the Google Play Store must be installed and the device must be
   /// running **Android 5 Lollipop(API 21)** or higher.
   ///
-  /// IOS devices must be running **IOS version 10.3** or higher.
+  /// iOS devices must be running **iOS version 10.3** or higher.
   ///
   /// MacOS devices must be running **MacOS version 10.14** or higher
   Future<bool> isAvailable() {
@@ -41,8 +41,8 @@ abstract class InAppReviewPlatform extends PlatformInterface {
   /// this cannot be done via [isAvailable]. If it is not available then
   /// you can open the store listing via [openStoreListing].
   ///
-  /// To improve the users experience, IOS and Android enforce limitations
-  /// that might prevent this from working after a few tries. IOS & MacOS users
+  /// To improve the users experience, iOS and Android enforce limitations
+  /// that might prevent this from working after a few tries. iOS & MacOS users
   /// can also disable this feature entirely in the App Store settings.
   ///
   /// More info and guidance:
@@ -54,13 +54,13 @@ abstract class InAppReviewPlatform extends PlatformInterface {
   }
 
   /// Opens the Play Store on Android, the App Store with a review
-  /// screen on IOS & MacOS and the Microsoft Store on Windows.
+  /// screen on iOS & MacOS and the Microsoft Store on Windows.
   ///
-  /// [appStoreId] is required for IOS & MacOS.
+  /// [appStoreId] is required for iOS & MacOS.
   ///
   /// [microsoftStoreId] is required for Windows.
   Future<void> openStoreListing({
-    /// Required for IOS & MacOS.
+    /// Required for iOS & MacOS.
     String? appStoreId,
 
     /// Required for Windows.