+++ in_app_review_platform_interface/lib/method_channel_in_app_review.dart
@@ -34,12 +34,12 @@ class MethodChannelInAppReview extends InAppReviewPlatform {
     String? appStoreId,
     String? microsoftStoreId,
   }) async {
-    final bool isIOS = _platform.isIOS;
+    final bool isiOS = _platform.isIOS;
     final bool isMacOS = _platform.isMacOS;
     final bool isAndroid = _platform.isAndroid;
     final bool isWindows = _platform.isWindows;
 
-    if (isIOS || isMacOS) {
+    if (isiOS || isMacOS) {
       await _channel.invokeMethod(
         'openStoreListing',
         ArgumentError.checkNotNull(appStoreId, 'appStoreId'),