+++ in_app_review_platform_interface/test/method_channel_in_app_review_test.dart
@@ -80,7 +80,7 @@ void main() {
       },
     );
     test(
-      'should invoke the openStoreListing method channel on IOS',
+      'should invoke the openStoreListing method channel on iOS',
       () async {
         // ARRANGE
         methodChannelInAppReview.platform =